# openapi.yaml
openapi: 3.0.3
info:
  title: BooklyX Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for BooklyX backend.

    Notes:
    - Platform-aware endpoints require `platform` header: `APP` or `WEB`.
    - Localized messages use `Accept-Language`: `en` or `ar`.
    - Protected endpoints require `Authorization: Bearer <token>`.
    - Access token format is `<loginSequence>|<jwt>` (example: `15|eyJ...`).
    - In development, OTP is commonly `333333`.
    - Response payload keys are returned in `snake_case`.
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://booklyxbackend-production.up.railway.app
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AcceptLanguageHeader:
      name: Accept-Language
      in: header
      required: false
      description: Preferred response language.
      schema:
        type: string
        enum: [en, ar]
        default: en
        example: en
      examples:
        english:
          summary: English language
          value: en
        arabic:
          summary: Arabic language
          value: ar
    PlatformHeader:
      name: platform
      in: header
      required: true
      description: Platform where request is coming from.
      schema:
        type: string
        enum: [APP, WEB]
        default: WEB
        example: WEB
      examples:
        app:
          summary: Mobile app client
          value: APP
        web:
          summary: Web client
          value: WEB
...
  schemas:
    ApiSuccess:
      type: object
      required: [status, error, message, data]
      properties:
        status:
          type: integer
          example: 200
        error:
          type: boolean
          example: false
        message:
          type: string
          example: Request successful.
        data:
          nullable: true
    ApiError:
      type: object
      required: [status, error, message, data]
      properties:
        status:
          type: integer
          example: 400
        error:
          type: boolean
          example: true
        message:
          type: string
          example: Invalid request.
        data:
          nullable: true
          example: null
        params:
          nullable: true
          example: null
...
  examples:
...
    StaffCompleteAppointmentRequest:
      value:
        notes: Client requested layered finish.
        attachments:
          - https://cdn.booklyx.com/executions/result-1.jpg
...
    StaffUpdateAvailabilityRequest:
      value:
        dayOfWeek: 2
        startTime: "10:00"
        endTime: "18:00"
...
    StaffAppointmentStartedResponse:
      value:
        status: 200
        error: false
        message: Appointment started successfully.
        data:
          id: 3
          status: IN_PROGRESS
    StaffAppointmentCompletedResponse:
      value:
        status: 200
        error: false
        message: Appointment completed successfully.
        data:
          id: 3
          status: COMPLETED
...
    StaffAvailabilityUpdatedResponse:
      value:
        status: 200
        error: false
        message: Availability slot updated successfully.
        data:
          id: 3
          dayOfWeek: 5
          startTime: "13:00"
          endTime: "21:00"
    StaffAvailabilityDeletedResponse:
      value:
        status: 200
        error: false
        message: Availability slot deleted successfully.
        data:
          id: 3
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
paths:
...
  /staff/appointments/{appointmentId}/start:
    patch:
      tags: [Staff/appointments]
      summary: Mark an appointment as in progress
      description: Mark an appointment as in progress.
      operationId: startAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/AcceptLanguageHeader"
        - $ref: "#/components/parameters/PlatformHeader"
      responses:
        "200":
          description: Appointment started
          content:
            application/json:
              examples:
                default:
                  $ref: "#/components/examples/StaffAppointmentStartedResponse"
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /staff/appointments/{appointmentId}/complete:
    patch:
      tags: [Staff/appointments]
      summary: Mark an appointment as completed
      description: Mark an appointment as completed.
      operationId: completeAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
        - $ref: "#/components/parameters/AcceptLanguageHeader"
        - $ref: "#/components/parameters/PlatformHeader"
      requestBody:
        required: false
        content:
          application/json:
            examples:
              default:
                $ref: "#/components/examples/StaffCompleteAppointmentRequest"
            schema:
              type: object
              properties:
                notes:
                  type: string
                  nullable: true
                attachments:
                  type: array
                  nullable: true
                  items:
                    type: string
                    format: uri
      responses:
        "200":
          description: Appointment completed
          content:
            application/json:
              examples:
                default:
                  $ref: "#/components/examples/StaffAppointmentCompletedResponse"
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /staff/availability/{availabilityId}:
    put:
      tags: [Staff/Availability]
      summary: Update a staff availability slot
      description: Update a staff availability slot. `dayOfWeek` uses JavaScript/dayjs numbering where `0 = Sunday`, `1 = Monday`, `2 = Tuesday`, `3 = Wednesday`, `4 = Thursday`, `5 = Friday`, and `6 = Saturday`.
      operationId: updateStaffAvailability
      security:
        - bearerAuth: []
      parameters:
        - name: availabilityId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - $ref: "#/components/parameters/AcceptLanguageHeader"
        - $ref: "#/components/parameters/PlatformHeader"
      requestBody:
        required: true
        content:
          application/json:
            examples:
              default:
                $ref: "#/components/examples/StaffUpdateAvailabilityRequest"
            schema:
              type: object
              properties:
                dayOfWeek:
                  type: integer
                  minimum: 0
                  maximum: 6
                  description: JavaScript/dayjs day index. 0 = Sunday, 1 = Monday, 2 = Tuesday, 3 = Wednesday, 4 = Thursday, 5 = Friday, 6 = Saturday.
                startTime:
                  type: string
                endTime:
                  type: string
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              examples:
                default:
                  $ref: "#/components/examples/StaffAvailabilityUpdatedResponse"
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Staff/Availability]
      summary: Delete a staff availability slot
      description: Delete a staff availability slot.
      operationId: deleteStaffAvailability
      security:
        - bearerAuth: []
      parameters:
        - name: availabilityId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - $ref: "#/components/parameters/AcceptLanguageHeader"
        - $ref: "#/components/parameters/PlatformHeader"
      responses:
        "200":
          description: Availability deleted
          content:
            application/json:
              examples:
                default:
                  $ref: "#/components/examples/StaffAvailabilityDeletedResponse"
              schema:
                $ref: "#/components/schemas/ApiSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...